[추가] : 에러 처리 추가
null값을 입력하였을 때, 이미 존재하는 아이디를 입력하였을 때
error처리 추가
