+++ src/main/java/com/codesquad/qna/advice/exception/InputMistakeException.java
@@ -1,0 +1,11 @@
+package com.codesquad.qna.advice.exception;
+
+public class InputMistakeException extends RuntimeException {
+    public InputMistakeException(String message) {
+        super(message);
+    }
+
+    public static InputMistakeException enteredNull() {
+        return new InputMistakeException("Null 값을 입력하였습니다");
+    }
+}
