+++ src/main/java/com/codesquad/qna/advice/exception/UnauthorizedException.java
@@ -1,4 +1,4 @@
-package com.codesquad.qna.exception.exception;
+package com.codesquad.qna.advice.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
@@ -13,11 +13,11 @@ public class UnauthorizedException extends RuntimeException {
         return new UnauthorizedException("로그인 해주세요");
     }
 
-    public static UnauthorizedException noMatchUser() {
-        return new UnauthorizedException("일치하는 사용자가 존재하지 않습니다");
+    public static InputMistakeException noMatchUser() {
+        return new InputMistakeException("일치하는 사용자가 존재하지 않습니다");
     }
 
-    public static UnauthorizedException noMatchPassword() {
-        return new UnauthorizedException("비밀번호가 일치하지 않습니다");
+    public static InputMistakeException noMatchPassword() {
+        return new InputMistakeException("비밀번호가 일치하지 않습니다");
     }
 }
